+++ lib/core/fl_theme.dart
@@ -8,9 +8,6 @@ class FLTheme {
     subtitle2: TextStyle(
       fontFamily: "RobotoMono",
     ),
-    button: TextStyle(
-      color: Colors.white,
-    ),
   );
 
   static final ThemeData lightTheme = ThemeData(